extract utility functions and higher level component
