+++ apps/newsletters-ui/src/app/components/RawDataDialog.tsx
@@ -13,43 +13,12 @@ import {
 	TableRow,
 } from '@mui/material';
 import { useState } from 'react';
-import type { ReactNode } from 'react';
+import { propertyToNode } from '../render-newsletter-properties';
 
 interface Props {
 	record: Record<string, unknown>;
 	title?: string;
 }
-
-const propertyToNode = (
-	record: Record<string, unknown>,
-	key: string,
-): ReactNode => {
-	const value = record[key];
-
-	switch (typeof value) {
-		case 'string':
-		case 'number':
-		case 'bigint':
-		case 'boolean':
-		case 'symbol':
-			return value.toString();
-		case 'undefined':
-			return '[UNDEFINED]';
-		case 'object':
-			try {
-				const stringification = JSON.stringify(value, undefined, 2);
-				return (
-					<pre style={{ fontSize: '75%', whiteSpace: 'pre-wrap' }}>
-						{stringification}
-					</pre>
-				);
-			} catch (err) {
-				return '[non-serialisable object]';
-			}
-		case 'function':
-			return `[function: ${value.name}]`;
-	}
-};
 
 export const RawDataDialog = ({ record, title = 'raw data' }: Props) => {
 	const [showRawData, setShowRawData] = useState(false);
@@ -94,7 +63,7 @@ export const RawDataDialog = ({ record, title = 'raw data' }: Props) => {
 									<TableCell size="small" sx={{ fontWeight: 'bold' }}>
 										{key}
 									</TableCell>
-									<TableCell>{propertyToNode(record, key)}</TableCell>
+									<TableCell>{propertyToNode(record[key])}</TableCell>
 								</TableRow>
 							))}
 						</TableBody>
